+++ ajax/insert-sell-return.php
@@ -6,26 +6,32 @@
 			$db = new Db();
 			  session_start();
 			  $datass = json_decode($_GET['dclas']);
+
 			  //delete before update 
-			  $mydate = $db->joinQuery("SELECT COUNT(*) as rowcount FROM `sell_return` WHERE `memono`='".$datass[0]->memo."'")->fetch(PDO::FETCH_ASSOC);
-			  if ($mydate['rowcount']>0) {
-			 $db->delete("sell_return","memono = '".$datass[0]->memo."'");
+			  if (isset($_GET['update']) && $_GET['update'] == true) 
+			  {
+			  	$mydate = $db->joinQuery("SELECT COUNT(*) as rowcount FROM `sell_return` WHERE `memono`='".$datass[0]->memo."'")->fetch(PDO::FETCH_ASSOC);
+					  if ($mydate['rowcount']>0) {
+					 $db->delete("sell_return","memono = '".$datass[0]->memo."'");
+					  }
 			  }
+			  
 			  //end of updating things
    					$cont = 0;
-			   for ($i=0; $i <count($datass); $i++) { 
+			   for ($i=0; $i <count($datass); $i++) 
+			   { 
 			   	
 			    	 $data = array(
 			    	 	
-			    	 	'memono' => $datass[$i]->memo,
-			    	 	'productid' => $datass[$i]->prod,
-			    	 	'customerid' => $_GET['custo'],
-			    	 	'quntity' => $datass[$i]->quntity,
-			    	 	'price' => $datass[$i]->price,
-			    	 	'weight' => $_GET['weight'],
-			    	 	'transport' => $_GET['transport'],
-			    	 	'return_date' => $datass[$i]->date,
-			    	 	'takenby' => $_SESSION['u_id']
+			    	 	'memono' => trim($datass[$i]->memo),
+			    	 	'productid' => trim($datass[$i]->prod),
+			    	 	'customerid' => trim($_GET['custo']) ,
+			    	 	'quntity' => trim($datass[$i]->quntity),
+			    	 	'price' => trim($datass[$i]->price),
+			    	 	'weight' => trim($_GET['weight']) ,
+			    	 	'transport' => trim($_GET['transport']),
+			    	 	'return_date' => trim($datass[$i]->date),
+			    	 	'takenby' => trim($_SESSION['u_id'])
 			    	 	 );
 
 			    	/* echo "<pre>";
@@ -37,12 +43,20 @@
 			    	
 			    	  
 			    }
+			    $msg = [];
 
-			    if (count($datass) == $cont) {
-			    	echo "Product has been returned";
-			    }else{
-			    	echo "There has been a problem";
+			    if (count($datass) == $cont) 
+			    {
+			    	array_push($msg, ["success"=>"Product has been returned"]);
+			    	//echo "Product has been returned";
 			    }
+			    else
+			    {
+			    	array_push($msg, ["error"=>"Problem occured"]);
+
+			    	//echo "There has been a problem";
+			    }
+			    echo json_encode($msg);
 			  /*echo "<pre>";
 			   print_r($_GET['allotherinfo']);
 			  print_r($_GET);
